requirements.txt
google-genai==1.24.0
customtkinter==5.2.2